requirements.txt
Flask
pdf2image
Pillow
Werkzeug
razorpay
pillow
reportlab
setuptools
wheel
fpdf2